apagar post autor e admin no feed
