+++ src/app/components/screens/Home/Comment.tsx
@@ -8,7 +8,7 @@ import InlineView from "../../reusable/InlineView"
 import { VariableHeightTextInput } from "../../reusable/VariableHeightTextInput"
 import { Button } from "../../reusable/Button"
 import { Text } from "../../reusable/Text"
-import { getCommentReplies, postComment } from "../../../lib/firebase/models"
+import { getCommentReplies, postComment } from "../../../lib/firebase/posts"
 import { UserContext } from "../../../providers/User"
 import { IconButton } from "react-native-paper"
 import { getUser } from "../../../lib/firebase/user"
