+++ src/app/components/screens/Home/PostComments.tsx
@@ -5,7 +5,7 @@ import { RootStackParamPostNav } from "./Post"
 import { VariableHeightTextInput } from "../../reusable/VariableHeightTextInput"
 import InlineView from "../../reusable/InlineView"
 import { Button } from "../../reusable/Button"
-import { getRootComments, postComment } from "../../../lib/firebase/models"
+import { getRootComments, postComment } from "../../../lib/firebase/posts"
 import { UserContext } from "../../../providers/User"
 import { FlatList } from "react-native"
 import Comment from "./Comment"
